pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-code-security-auditor"
version = "2.0.0"
description = "AI-powered code security scanner with CLI and API support"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "AI Security Team" }
]
dependencies = [
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "python-dotenv>=1.0.0",
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "prometheus-client>=0.20.0",
  "sqlalchemy>=2.0.0",
  "celery>=5.3.0",
  "redis>=5.0.0",
  "openai>=1.0.0",
  "httpx>=0.25.0",
  "langgraph>=0.0.40",
  "langchain>=0.1.0",
  "langchain-core>=0.1.0",
  "bandit>=1.7.5",
  "semgrep>=1.45.0",
  "click>=8.1.7",
  "rich>=13.0.0",
  "colorama>=0.4.6",
  "tqdm>=4.65.0",
  "chromadb>=0.4.15",
  "sentence-transformers>=2.2.2"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Security",
  "Topic :: Software Development :: Quality Assurance"
]

[project.scripts]
auditor = "auditor.cli:main"
ai-security-auditor = "auditor.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*", "auditor*", "cli_visuals*"]
exclude = ["tests*", "scripts*", "chroma_db*", "venv*"]
